Complete user landing page query
